Refactored ServiceRequest to simplify the code and leverage more features of the standard library. For example, there was no need to construct the URL in such a manual way. Also made a number of methods static, which is cleaner and will make them easier to test.
Added a feature where method parameters may be passed directly to call_service() instead of setting them on the service object. This will allow a method to be called with just one line of code instead of a line to create the service object, then another line for each parameter, and a line to finally make the call. A call to a 3-parameter method previously took 5 lines of code, and will now take only 1.

Overall there is less code, it is more useable, and it is easier to test. If we could get rid of the idea that a ServiceRequest has a 'parameters' attribute at all (which would be a great improvement), we could remove even more code. However, I've left that feature in for backward-compatibility.